+++ SpringBoot/src/main/java/com/seungwook/ktsp/global/auth/controller/EmailVerifyController.java
@@ -1,7 +1,7 @@
 package com.seungwook.ktsp.global.auth.controller;
 
 import com.seungwook.ktsp.global.auth.dto.request.VerifyRequest;
-import com.seungwook.ktsp.global.auth.service.AuthService;
+import com.seungwook.ktsp.global.auth.service.VerificationService;
 import com.seungwook.ktsp.global.response.Response;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -13,13 +13,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-@Tag(name = "Auth", description = "인증/인가 관련 API")
+@Tag(name = "Email Verification", description = "이메일 인증 관련 API")
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/auth")
+@RequestMapping("/verify")
 public class EmailVerifyController {
 
-    private final AuthService authService;
+    private final VerificationService verificationService;
 
     // 인증코드 발송
     @Operation(summary = "이메일 인증코드 발송", description = "해당 이메일로 인증코드가 발송됨")
@@ -31,7 +31,7 @@ public class EmailVerifyController {
 
         String userEmail = email + "@kangwon.ac.kr";
 
-        authService.sendAuthCode(userEmail);
+        verificationService.sendAuthCode(userEmail);
 
         return ResponseEntity.ok(Response.<Void>builder()
                 .message("인증코드 메일 발송 성공")
@@ -40,11 +40,11 @@ public class EmailVerifyController {
 
     // 인증코드 검증
     @Operation(summary = "이메일 인증코드 검증", description = "인증코드가 유효한지 검증")
-    @PostMapping("/verify")
+    @PostMapping
     public ResponseEntity<Response<Void>> verify(@Valid @RequestBody VerifyRequest request) {
         String userEmail = request.getEmail() + "@kangwon.ac.kr";
 
-        authService.verifyAuthCode(userEmail, request.getVerifyCode().toUpperCase());
+        verificationService.verifyAuthCode(userEmail, request.getVerifyCode().toUpperCase());
 
         return ResponseEntity.ok(Response.<Void>builder()
                 .message("인증 성공")
