+++ SpringBoot/src/main/java/com/seungwook/ktsp/global/auth/exception/UserContextException.java
@@ -1,0 +1,11 @@
+package com.seungwook.ktsp.global.auth.exception;
+
+import com.seungwook.ktsp.global.exception.BaseCustomException;
+import org.springframework.http.HttpStatus;
+
+public class UserContextException extends BaseCustomException {
+
+    public UserContextException(String message) {
+        super(HttpStatus.UNAUTHORIZED, message);
+    }
+}
